Handle processor indices saved by BOUT++
v4.3 of BOUT++ saves processor indices to dump files as 'PE_XIND' and
'PE_YIND'. Use these variables to identify boundary cells if they are
present, falling back on the number in the filename if they are not.
Trim 'PE_XIND' and 'PE_YIND' before concatenating, since they are not
the same across processors.