Message and workaround when name of coordinate conflicts with variable
When adding physical coordinates ('geometry') to a dataset, it is an
error if the name of one of the coordinates is the same as the name of a
variable in the dataset. Provide informative error messages if this
happens. Give a workaround - now there is a 'coordinates' argument to
open_boutdataset (and various lower-level functions) that can be used to
pass a tuple/list of names to be used instead of the defaults.