Enable reading of single, squashed output files
If a single output file (or a series of single files from consecutive
runs) is opened with xBOUT, the load would previously fail because
NXPE*NYPE>1, while only one file is found. This commit introduces some
special handling for the case when all x- and y-points are found in the
file, setting NXPE and NYPE to 1 instead of the value in the files, and
setting mxg and myg to 0 if boundary cells in x- or y-directions are not
found in the file.