Don't store _grid in dataset attrs
Doing this prevented the dataset being saved to netCDF file.

Also use _open_grid instead of open_boutdataset to open the grid file so
that the grid dataset does not have metadata added to the DataArray
variables. These variables are added as coordinates and become members
of the DataArrays representing simulation variables; if they have a
metadata dict, it is not possible to save them to netCDF.