Don't try to store grid dataset after reading coordinates
Previously, a gridfile was loaded and stored as an attribute of the
Dataset and DataArrays, then it was attempted to store the grid Dataset
as a member of the BoutDataset accessor (this does not work because
accessors are not permanent).

As of BOUT++ v4.3.0, the only information we need to load from the grid
file is some coordinates (e.g. psixy, Rxy, Zxy for 'toroidal' geometry),
so we can just open the grid file, set the coordinates and not save the
grid. This has the advantage that DataArrays inherit the coordinates
from their parent Dataset without having to do special things when
loading, or causing problems when saving to netCDF.