Ensure separatrices, target plotted correctly if Dataset transposed
Create metadata variables called 'bout_tdim', 'bout_xdim', 'bout_ydim',
and 'bout_zdim' that record what 't', 'x', 'y', and 'z' respectively
were renamed to. This means that in plot_separatrices() and
plot_targets() we can find renamed dimensions corresponding to 'x' and
'y' even if the Dataset has been transposed so that we cannot rely on
the order of the dimensions.