Arguments to make logarithmic color-scale in 2d plots
'norm' option allows any matplotlib.colors.Normalize instance to be
passed, for general control of the color-scale.

For convenience, also adds a 'logscale' option which can be set to True
to create a standard log-scale if vmin and vmax are both positive or
both negative, and a symmetric log-scale with a linear threshold of
min(abs(vmin),abs(vmax))*1.e-5 if vmin and vmax have opposite signs. If
a float is passed for 'logscale' and vmin and vmax have opposite signs
then the value of 'logscale' is used instead of the default 1.e-5 to set
the linear threshold.