More robust handling of 'chunks' when opening grid files
Grid files don't have a 't' dimension, and usually don't have a 'z'
dimension, so if either are passed in the 'chunks' argument, previously
there would have been an error. This commit checks the dimensions in
'chunks' passed to a grid file are only 'x', 'y', or 'z', and drops 'z'
if it is passed and is not present in the grid file.