Fix coordinates of guard cells added to a region to make them continuous
Guard cells taken from a neighbouring region may have coordinates that
are not continuous with the coordinates of the region they are being
added to. Work around this by replacing the coordinate values of the
guard cells being added with the coordinate values of the adjacent
couple of points in the global grid (which have values continuous with
those in the region being added to).

Adds some utility methods to Region to get slices to select the values
to fill the guard cells with.