Allow geometries to be re-applied to a Dataset
After interpolating to higher parallel resolution, a Dataset has the
correct coordinates, but no regions. This commit makes
add_toroidal_geometry_coords and add_s_alpha_geometry_coords skip
adding coordinates if the coordinates are already present, so that the
functions can be applied again to interpolated Datasets. At the moment,
the only thing this does is to re-create the regions.