Raise exception if create_bout_ds_list is called with nt>1
If nt>1 then create_bout_ds_list should create nt subdirectories of data
(each from a different set of times), but this is not implemented yet.
Previously, if _bout_xyt_example_files was called with nt>1 there would
be an error, because create_bout_ds_list did not create subdirectories,
but the glob pattern in _bout_xyt_example_files expected subdirectories,
with the side effect that the glob included all the temporary data*
directories created in the test. When running a single test this would
not cause an error, but when running multiple tests that create
different directories it might.