Add pytest.mark.long to disable long tests by default
Mark a test as 'long' by using the mark 'pytest.mark.long'. 'long' tests
are skipped by default, and can be enabled by passing the option
'--long' on the command line. Long tests are enabled on Travis.

Most of the combinations of inputs for the Region and parallel
interpolation tests are marked as long, to speed up pytest runs.