Prefer dz to ZMAX and ZMIN
When calculating the size of the toroidal domain, prefer 'dz' to 'ZMAX'
and 'ZMIN'. 'dz' is the variable actually used for derivatives in
BOUT++, so while 'dz' and '2*pi*(ZMAX - ZMIN)' should be consistent,
'dz' is more certain to be consistent with the simulation data.

Also fixes the calculation of the toroidal angle coordinate, where the
np.linspace() should have had 'endpoint=False' passed since the last
point on the grid is adjacent to the first one, not in the same place.