Only one region in regions attr of result of BoutDataArray.from_region()
Instead of adding a 'region' attr to the result of
BoutDataArray.from_region(), replace the 'regions' attr with a dict
containing only the region of the returned BoutDataArray. Allows
poloidal plots like
"ds['n'].bout.from_region('upper_outer_SOL').bout.pcolomesh()" which
show only the variable in a single region. Also, the single region left
in attrs['regions'] has all the boundary indices set to None, so that
from_region() be called again on the BoutDataArray inside plotting
methods.