interpolate_from_unstructured sets all points outside grid to fill_value
Previously, any points within the 'convex hull' of the grid were
interpolated to, even if they were outside the actual grid - this was
because of the way scipy.interpolate.griddata works. Now use
matplotlib.path.Path.contains_points() to use the actual boundary of the
grid and set everything outside of that to 'fill_value'.