Check dimensions of variables to interpolate
Especially when passing '...' to interpolate all variables, need to
check wether variable in the variable list has both x- and y-dimenions,
otherwise it can't be interpolated - and if it has no y-dimension,
doesn't need to be interpolated.

Also ensure the first variable added to the high-resolution Dataset has
all dimensions of the original Dataset (or at least as many as possible)
so that all coordinates get passed on to the high-resolution Dataset.