Better support for generic x and y in poloidal plots
By default, the x and y coordinates for poloidal plots are "R" and "Z",
but they are supposed to be allowed to be set to any 2d coordinates on
the poloidal plane. This commit adds support for generic x and y to
plot_separatrices, plot_targets and to the X-point region check in
plot2d_wrapper.