Add reload_boutdataset()
New function reload_boutdataset() loads a BoutDataset from a netCDF file
created by `ds.bout.save()`. Necessary because for example dimensions
may be renamed when adding geometry, and metadata must be converted from
dict to attributes to be saved, so reloading the saved Dataset is not
the same as initial loading from BOUT++ output files.