Better fix, BoutDataArray.interpolate_parallel() performance regression
Removing attrs from y-coordinate means we do not need to call
da.compute(), which would load the entire result into memory. It is
better not to, as the result may be sliced or processed somehow later
and we don't want to force loading in case the variable is too large to
fit in memory.