Work around xarray bug; only add coordinates if not already present
An xarray bug
(https://github.com/pydata/xarray/issues/2180#issuecomment-391932929,
https://github.com/pydata/xarray/issues/4393,
https://github.com/pydata/xarray/issues/4415) can deletes attributes of
coordinates when a new DataArray is assigned to a Dataset. This commit
adds workarounds for that bug in several places.

Also adds checks for if 'ycoord' and 'zcoord' already exist in
apply_geometry() before creating them. Helps with re-applying
apply_geometry() to existing or re-loaded Datasets.