Work around xarray using scipy engine to write netcdf
If a 'LocalPath' as created by pytest is passed as the argument to
xarray's to_netcdf() method, it uses the scipy engine because it thinks
the argument is a 'file object'. To avoid this convert the path to a
str. We want to avoid the scipy engine because it produces a
DeprecationWarning.