Change default of matplotlib.rcParams["pcolor.shading"] to "auto"
This is a new parameter in matplotlib-3.3. The default is the
backward-compatible is "flat" which drops the last row and column of C
if X, Y and C have the same size. The new "auto" uses the X, Y values as
grid point values if X, Y and C are all the same size and as cell
corner values if X and Y are one point bigger than C.