Fix multiple-coordinate-values error for core-only/limiter cases
When values of the theta coordinate for the guard cells were taken from
the other side of the branch cut for core-only or limiter cases, they
did not join continuously onto the main grid's theta values. Work around
this by adding an offset calculated from a linear extrapolation from the
grid values (which is exact for constant grid spacing, which is usually
the case for the y-direction).