Allow geometry to be changed when reloading an xBOUT-saved Dataset
If a Dataset is loaded with xBOUT using one geometry, and then saved
(e.g. using geometry=None just to squash the data into a single file),
and then reloaded, previously the geometry was re-applied from whatever
was set at the first load. Now it is possible to override the geometry
while reloading (e.g. to apply toroidal geometry when opening the
squashed file).