Don't set coordinates in create_bout_ds()
Don't want to save coordinates into test data files, because this does
not replicate the structure of actual BOUT++ output files - in
particular if there were coordinates in the files, restart files in the
tests get coordinates because the coordinates are saved in `_encoding`
of Variables. Requires some changes in handling of coordinates in the
tests.