Allow loading restart files
A few small work-arounds are needed to load restart files and to save
new restart files from a Dataset loaded from restart files, because
restart files have no t-dimension and may be missing some variables that
are present in dump files. By default, we detect restart files by
checking if `restart` is in the file name, but the work-arounds can be
forced on or off by passing `True` or `False` to the `is_restart`
argument to `open_boutdataset()`.