Work around NaN in corner cells of zShift
When zShift is saved by BOUT++ it may have NaN in some corner
boundary/guard cells. When doing FFTs of arrays, these NaNs can make the
whole output NaN, even in entries that should not depend on the corner
cells of zShift. To avoid this, replace the NaNs with 0 - nothing
important should depend on the corner cell values, so this should not
cause problems.