Fix handling of 'last output step' metadata variables
Handle 'hist_hi' and 'tt' like 'iteration'. Note 'iteration' actually
doesn't make as much sense, because it is related to the number of
iterations in each individual run, so taking the value from the latest
run doesn't really make sense for 'iteration'. It does make sense for
'hist_hi' and 'tt' - the values of those would be the same when opening
the output from multiple, restarted runs as they would when opening a
output from an equivalent single run (with no restarts).

Also, move the handling of `_BOUT_TIME_DEPENDENT_META_VARS` to a point
before metadata is split, so that the handled variables get added to
metadata in the same way regardless of whether one restart or multiple
restarts are loaded.