Workaround for old squashed files with inconsistent y-boundary cells
Older 'squashed' output files for double-null grids often included
y-boundary cells for the lower target (ends of the logical grid) but not
the upper target (somewhere in the middle of the logical grid). Add some
special handling to allow these files to be opened, as long as
`keep_yboundaries = False`.